Move dependency lifecycle management to separate modules
