+++ apps/desktop/src-tauri/src/main.rs
@@ -3,15 +3,15 @@
     windows_subsystem = "windows"
 )]
 
+mod backend;
 mod qdrant;
 
 use std::path::PathBuf;
 
-use bleep::{Application, Configuration, Environment};
+use bleep::Application;
 use once_cell::sync::OnceCell;
 use sentry::ClientInitGuard;
 use std::sync::{Arc, RwLock};
-use tauri::Manager;
 
 static TELEMETRY: RwLock<bool> = RwLock::new(false);
 static SENTRY: OnceCell<ClientInitGuard> = OnceCell::new();
@@ -41,61 +41,8 @@ async fn main() {
     Application::install_logging();
 
     tauri::Builder::default()
-        .setup(|app| {
-            let config = app
-                .path_resolver()
-                .resolve_resource("config.json")
-                .expect("failed to resolve resource");
-
-            let mut configuration = Configuration::read(config).unwrap();
-            configuration.ctags_path = relative_command_path("ctags");
-            configuration.max_threads = bleep::default_parallelism() / 4;
-            configuration.model_dir = app
-                .path_resolver()
-                .resolve_resource("model")
-                .expect("bad bundle");
-
-            let cache_dir = app.path_resolver().app_cache_dir().unwrap();
-            configuration
-                .source
-                .set_default_dir(&cache_dir.join("bleep"));
-
-            tokio::task::block_in_place(move || {
-                tokio::runtime::Handle::current().block_on(qdrant::start(cache_dir))
-            });
-
-            let app = app.handle();
-            tokio::spawn(async move {
-                let initialized =
-                    Application::initialize(Environment::InsecureLocal, configuration).await;
-
-                if let Ok(backend) = initialized {
-                    if let Err(_e) = backend.run().await {
-                        app.emit_all(
-                            "server-crashed",
-                            Payload {
-                                message: _e.to_string(),
-                            },
-                        )
-                        .unwrap()
-                    }
-                } else {
-                    app.emit_all(
-                        "server-crashed",
-                        Payload {
-                            message: "Something bad happened".into(),
-                        },
-                    )
-                    .unwrap();
-                }
-            });
-
-            Ok(())
-        })
-        .on_window_event(|window_event| match window_event.event() {
-            tauri::WindowEvent::CloseRequested { .. } => qdrant::shutdown(),
-            _ => {}
-        })
+        .plugin(qdrant::QdrantSupervisor::default())
+        .setup(backend::bleep)
         .invoke_handler(tauri::generate_handler![
             show_folder_in_finder,
             get_device_id,
